Splitted control_node.cpp into control_node.cpp, robot_layer.cpp and robot_layer.h
