+++ canopen_motor_node/CMakeLists.txt
@@ -101,6 +101,7 @@ catkin_package(
 ## Your package locations should be listed before other locations
 # include_directories(include)
 include_directories(
+  include
   ${catkin_INCLUDE_DIRS}
   ${MUPARSER_INCLUDE_DIRS}
 )
@@ -111,7 +112,9 @@ include_directories(
 # )
 
 ## Declare a cpp executable
-add_executable(canopen_motor_node src/control_node.cpp)
+add_executable(canopen_motor_node
+  src/robot_layer.cpp
+  src/control_node.cpp)
 
 ## Add cmake target dependencies of the executable/library
 ## as an example, message headers may need to be generated before nodes
